ref: swapped `MoveService` and `CharacterService` functions
`MoveService` now saves the game object position and broadcasts
`CharacterService` now saves the location to our `Character` type
